+++ ifinitySDK.framework/Versions/A/Headers/IfinitySDK.h
@@ -6,8 +6,8 @@
 //  Copyright (c) 2014 GetIfinity. All rights reserved.
 //
 
-#define IFINITY_SDK_VERSION_STRING @"0.1.2"
-#define IFINITY_SDK_VERSION 1002
+#define IFINITY_SDK_VERSION_STRING @"0.1.3"
+#define IFINITY_SDK_VERSION 1003
 
 
 #import "IFDataManager.h"
@@ -37,6 +37,7 @@
 #import "IFMBeacon.h"
 #import "IFMContent.h"
 #import "IFMFloorplan.h"
+#import "IFMVenue.h"
 
 
 extern int if_ifinity_sdk_version();